Flytter en del extension-funksjoner fra AndelTilkjentYtelse til ØkonomiUtils, der de hører hjemme
